requirement.txt
streamlit
google-genai
fastapi[all]
PyPDF2
logfire[fastapi]
pydantic-settings